+++ go.mod
@@ -2,8 +2,18 @@ module github.com/anccy/lottery
 
 go 1.18
 
+require github.com/spf13/cobra v1.6.1
+
 require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/go-lark/lark v1.7.4 // indirect
 	github.com/inconshreveable/mousetrap v1.1.0 // indirect
-	github.com/spf13/cobra v1.6.1 // indirect
+	github.com/nikoksr/notify v0.38.0 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
 	github.com/spf13/pflag v1.0.5 // indirect
+	github.com/stretchr/objx v0.5.0 // indirect
+	github.com/stretchr/testify v1.8.1 // indirect
+	golang.org/x/sync v0.1.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
